+++ sketch/src/main/java/com/github/panpf/sketch/util/Size.kt
@@ -47,6 +47,8 @@ open class Size(val width: Int, val height: Int) {
 
     companion object {
 
+        val Empty = Size(0, 0)
+
         @Throws(NumberFormatException::class)
         fun parseSize(string: String): Size {
             var sepIx = string.indexOf('*')