+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # new
 
+#### sketch
+
+* new: Size added Empty constant
+
 #### sketch-zoom
 
 * fix: Fixed a bug where SketchZoomImageView's auto-pause feature based on lifecycle could cause subsampling exceptions